move FakeLazyJson to conftest
