--- tests/fake_lazy_json.py
@@ -1,15 +1,0 @@
-from types import TracebackType
-from typing import Self
-
-
-class FakeLazyJson(dict):
-    def __enter__(self) -> Self:
-        return self
-
-    def __exit__(
-        self,
-        exc_type: type[BaseException] | None,
-        exc_val: BaseException | None,
-        exc_tb: TracebackType | None,
-    ) -> None:
-        pass
